# docs/swagger.yaml
basePath: /
definitions:
  handlers.AddSongRequest:
    properties:
      group:
        example: Queen
        type: string
      song:
        example: Bohemian Rhapsody
        type: string
    type: object
  models.Song:
    properties:
      created_at:
        type: string
      group:
        type: string
      id:
        type: integer
      link:
        type: string
      releaseDate:
        type: string
      song:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Это REST API для управления музыкальной библиотекой
  title: Онлайн-библиотека песен API
  version: "1.0"
paths:
  /songs:
    get:
      description: Возвращает список всех доступных песен
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Song'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список песен
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Добавляет новую песню в базу данных
      parameters:
      - description: Группа и название песни
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddSongRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Song'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить песню
      tags:
      - songs
  /songs/{id}:
    delete:
      description: Удаляет песню по ID
      parameters:
      - description: ID песни
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить песню
      tags:
      - songs
swagger: "2.0"